requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
black==22.3.0
cachetools==4.2.2
certifi==2021.10.8
click==8.1.2
flake8==4.0.1
greenlet==1.1.2
importlib-metadata==4.2.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
pycodestyle==2.8.0
pyflakes==2.4.0
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
SQLAlchemy==1.4.35
tomli==2.0.1
tornado==6.1
typed-ast==1.5.2
typing-extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
zipp==3.8.0
